Refactor: Rename and reorganize `spring-rest-multidb-mongo-redis-postgres` module to `spring-rest-polyglot-persistence`.
