+++ core-tests/e2e-tests/spring/pom.xml
@@ -26,7 +26,7 @@
         <module>spring-rest-bb</module>
         <module>spring-graphql-bb</module>
         <module>spring-rest-multidb</module>
-        <module>spring-rest-multidb-mongo-redis-postgres</module>
+        <module>spring-rest-polyglot-persistence</module>
         <module>spring-rest-opensearch</module>
         <module>spring-rest-redis</module>
         <module>spring-rest-rsa</module>
