--- core-tests/e2e-tests/spring/spring-rest-multidb-mongo-redis-postgres/src/main/kotlin/com/foo/spring/rest/multidb/mongoredispostgres/CombinedDataDto.kt
@@ -1,7 +1,0 @@
-package com.foo.spring.rest.multidb.mongoredispostgres
-
-data class CombinedDataDto(
-    val postgresFound: Boolean,
-    val mongoFound: Boolean,
-    val redisFound: Boolean
-)
