--- core-tests/e2e-tests/spring/spring-rest-multidb-mongo-redis-postgres/src/main/kotlin/com/foo/spring/rest/multidb/mongoredispostgres/MongoPerson.kt
@@ -1,15 +1,0 @@
-package com.foo.spring.rest.multidb.mongoredispostgres
-
-import javax.persistence.Id
-
-class MongoPerson {
-    @Id
-    var id: String? = null
-    var name: String? = null
-
-    constructor()
-
-    constructor(name: String) {
-        this.name = name
-    }
-}
