--- core-tests/e2e-tests/spring/spring-rest-multidb-mongo-redis-postgres/src/main/kotlin/com/foo/spring/rest/multidb/mongoredispostgres/MongoPersonRepository.kt
@@ -1,7 +1,0 @@
-package com.foo.spring.rest.multidb.mongoredispostgres
-
-import org.springframework.data.mongodb.repository.MongoRepository
-
-interface MongoPersonRepository : MongoRepository<MongoPerson, String> {
-    fun findByName(name: String): List<MongoPerson>
-}
