+++ core-tests/e2e-tests/spring/spring-rest-polyglot-persistence/src/main/kotlin/com/foo/spring/rest/polyglotpersistence/CombinedDataDto.kt
@@ -1,0 +1,7 @@
+package com.foo.spring.rest.polyglotpersistence
+
+data class CombinedDataDto(
+    val postgresFound: Boolean,
+    val mongoFound: Boolean,
+    val redisFound: Boolean
+)
