+++ core-tests/e2e-tests/spring/spring-rest-polyglot-persistence/src/main/kotlin/com/foo/spring/rest/polyglotpersistence/MongoPerson.kt
@@ -1,0 +1,15 @@
+package com.foo.spring.rest.polyglotpersistence
+
+import javax.persistence.Id
+
+class MongoPerson {
+    @Id
+    var id: String? = null
+    var name: String? = null
+
+    constructor()
+
+    constructor(name: String) {
+        this.name = name
+    }
+}
