+++ core-tests/e2e-tests/spring/spring-rest-polyglot-persistence/src/main/kotlin/com/foo/spring/rest/polyglotpersistence/MongoPersonRepository.kt
@@ -1,0 +1,7 @@
+package com.foo.spring.rest.polyglotpersistence
+
+import org.springframework.data.mongodb.repository.MongoRepository
+
+interface MongoPersonRepository : MongoRepository<MongoPerson, String> {
+    fun findByName(name: String): List<MongoPerson>
+}
