+++ core-tests/e2e-tests/spring/spring-rest-polyglot-persistence/src/main/resources/schema/V1.0__createDB.sql
